refactor(merchant): enhance data handling and type consistency
- Introduce `fromArray` method in `MerchantDto` for array-to-object conversion
- Replace string timestamps with `DateTimeInterface` for `createdAt` and `updatedAt`
- Add `castComissionRates` and `castDatetime` methods for data casting
- Simplify `MerchantResponse` by using `fromArray` method
- Update `MerchantDtoTest` to validate `DateTimeInterface` and `CommissionRate` instances
- Modify `CommissionRate` to use `float` for commission values instead of `int`

+++ src/Dto/V1/Monetize/V2/Merchant/CommissionRate.php
@@ -9,12 +9,12 @@ namespace Muscobytes\TakeAdsApi\Dto\V1\Monetize\V2\Merchant;
 final readonly class CommissionRate
 {
     /**
-     * @param int|null $fixedCommission Set amount paid per transaction or lead.
-     * @param int|null $percentageCommission Percentage of the total sale amount paid per transaction.
+     * @param float|null $fixedCommission Set amount paid per transaction or lead.
+     * @param float|null $percentageCommission Percentage of the total sale amount paid per transaction.
      */
     public function __construct(
-        public ?int $fixedCommission,
-        public ?int $percentageCommission
+        public ?float $fixedCommission,
+        public ?float $percentageCommission
     )
     {
         //
